refactor: event 패키지 변경
