+++ flight-service/src/main/java/com/oneul_tanda/flight_service/application/event/FlightEventKafkaListener.java
@@ -1,5 +1,7 @@
 package com.oneul_tanda.flight_service.application.event;
 
+import com.oneul_tanda.flight_service.application.dtos.event.ReservationCanceledEvent;
+
 public interface FlightEventKafkaListener {
 
     void consumePaymentCanceledFromPayment(ReservationCanceledEvent event);
